+++ engine/src/execution_graph/logic_controllers/BatchOrderByProcessing.cpp
@@ -15,13 +15,13 @@ PartitionSingleNodeKernel::PartitionSingleNodeKernel(std::size_t kernel_id, cons
 }
 
 void PartitionSingleNodeKernel::do_process(std::vector< std::unique_ptr<ral::frame::BlazingTable> > inputs,
-    std::shared_ptr<ral::cache::CacheMachine> output,
-    cudaStream_t stream, const std::map<std::string, std::string>& args) {
+    std::shared_ptr<ral::cache::CacheMachine> /*output*/,
+    cudaStream_t /*stream*/, const std::map<std::string, std::string>& /*args*/) {
     auto & input = inputs[0];
 
     auto partitions = ral::operators::partition_table(partitionPlan->toBlazingTableView(), input->toBlazingTableView(), this->expression);
 
-    for (auto i = 0; i < partitions.size(); i++) {
+    for (std::size_t i = 0; i < partitions.size(); i++) {
         std::string cache_id = "output_" + std::to_string(i);
         this->add_to_output_cache(
             std::make_unique<ral::frame::BlazingTable>(std::make_unique<cudf::table>(partitions[i]), input->names()),
@@ -86,16 +86,21 @@ SortAndSampleKernel::SortAndSampleKernel(std::size_t kernel_id, const std::strin
     this->query_graph = query_graph;
     set_number_of_message_trackers(2); //default
     this->output_.add_port("output_a", "output_b");
+    get_samples = true;
+    already_computed_partition_plan = false;
 }
 
-void SortAndSampleKernel::compute_partition_plan(std::vector<ral::frame::BlazingTableView> sampledTableViews, std::size_t avg_bytes_per_row, std::size_t local_total_num_rows) {
+void SortAndSampleKernel::compute_partition_plan(
+    std::vector<std::unique_ptr<ral::frame::BlazingTable>> inputSamples,
+    std::size_t avg_bytes_per_row,
+    std::size_t local_total_num_rows) {
 
     if (this->context->getAllNodes().size() == 1){ // single node mode
-        auto partitionPlan = ral::operators::generate_partition_plan(sampledTableViews,
+        auto partitionPlan = ral::operators::generate_partition_plan(inputSamples,
             local_total_num_rows, avg_bytes_per_row, this->expression, this->context.get());
         this->add_to_output_cache(std::move(partitionPlan), "output_b");
     } else { // distributed mode
-        if( ral::utilities::checkIfConcatenatingStringsWillOverflow(sampledTableViews)) {
+        if( ral::utilities::checkIfConcatenatingStringsWillOverflow(inputSamples)) {
             logger->warn("{query_id}|{step}|{substep}|{info}",
                             "query_id"_a=(context ? std::to_string(context->getContextToken()) : ""),
                             "step"_a=(context ? std::to_string(context->getQueryStep()) : ""),
@@ -111,7 +116,7 @@ void SortAndSampleKernel::compute_partition_plan(std::vector<ral::frame::Blazing
             std::vector<std::unique_ptr<ral::cache::CacheData> >table_scope_holder;
             std::vector<size_t> total_table_rows;
             std::vector<size_t> total_avg_bytes_per_row;
-            std::vector<ral::frame::BlazingTableView> samples;
+            std::vector<std::unique_ptr<ral::frame::BlazingTable>> samples;
 
             for(std::size_t i = 0; i < nodes.size(); ++i) {
                 if(!(nodes[i] == self_node)) {
@@ -122,12 +127,12 @@ void SortAndSampleKernel::compute_partition_plan(std::vector<ral::frame::Blazing
 
                     total_table_rows.push_back(std::stoll(cache_ptr->getMetadata().get_values()[ral::cache::TOTAL_TABLE_ROWS_METADATA_LABEL]));
                     total_avg_bytes_per_row.push_back(std::stoll(cache_ptr->getMetadata().get_values()[ral::cache::AVG_BYTES_PER_ROW_METADATA_LABEL]));
-                    samples.push_back(cache_ptr->getTableView());
+                    samples.push_back(cache_ptr->decache());
                 }
             }
 
-            for (std::size_t i = 0; i < sampledTableViews.size(); i++){
-                samples.push_back(sampledTableViews[i]);
+            for (std::size_t i = 0; i < inputSamples.size(); i++){
+                samples.push_back(std::move(inputSamples[i]));
             }
 
             total_table_rows.push_back(local_total_num_rows);
@@ -156,6 +161,11 @@ void SortAndSampleKernel::compute_partition_plan(std::vector<ral::frame::Blazing
         } else {
             context->incrementQuerySubstep();
 
+            // just to concat all the samples
+            std::vector<ral::frame::BlazingTableView> sampledTableViews;
+            for (std::size_t i = 0; i < inputSamples.size(); i++){
+                sampledTableViews.push_back(inputSamples[i]->toBlazingTableView());
+            }
             auto concatSamples = ral::utilities::concatTables(sampledTableViews);
             concatSamples->ensureOwnership();
 
@@ -179,112 +189,103 @@ void SortAndSampleKernel::compute_partition_plan(std::vector<ral::frame::Blazing
     }
 }
 
-kstatus SortAndSampleKernel::run() {
-    CodeTimer timer;
+void SortAndSampleKernel::do_process(std::vector< std::unique_ptr<ral::frame::BlazingTable> > inputs,
+    std::shared_ptr<ral::cache::CacheMachine> output,
+    cudaStream_t /*stream*/, const std::map<std::string, std::string>& args) {
+
     CodeTimer eventTimer(false);
 
-    bool try_num_rows_estimation = true;
-    bool get_samples = true;
-    bool estimate_samples = false;
-    uint64_t num_rows_estimate = 0;
-    uint64_t population_to_sample = 0;
-    uint64_t population_sampled = 0;
-    BlazingThread partition_plan_thread;
+    auto& operation_type = args.at("operation_type");
 
-    float order_by_samples_ratio = 0.1;
-    std::map<std::string, std::string> config_options = context->getConfigOptions();
-    auto it = config_options.find("ORDER_BY_SAMPLES_RATIO");
-    if (it != config_options.end()){
-        order_by_samples_ratio = std::stof(config_options["ORDER_BY_SAMPLES_RATIO"]);
+    auto & input = inputs[0];
+    if (operation_type == "ordering_and_get_samples") {
+        auto sortedTable = ral::operators::sort(input->toBlazingTableView(), this->expression);
+
+        if (get_samples.load()) { 
+            auto sampledTable = ral::operators::sample(input->toBlazingTableView(), this->expression);
+            
+            std::lock_guard<std::mutex> samples_lock(samples_mutex);
+            if (get_samples.load()) {
+                population_sampled += sampledTable->num_rows(); 
+                localTotalNumRows += input->view().num_rows();
+                localTotalBytes += input->sizeInBytes();
+                samplesTables.push_back(std::move(sampledTable));
+                if (population_sampled > max_order_by_samples) {
+
+                    std::vector<std::unique_ptr <ral::cache::CacheData> > sampleCacheDatas;
+                    for (std::size_t i = 0; i < samplesTables.size(); ++i) {
+                        std::unique_ptr <ral::cache::CacheData> cache_data = std::make_unique<ral::cache::GPUCacheData>(std::move(samplesTables[i]));
+                        sampleCacheDatas.push_back(std::move(cache_data));
+                    }
+
+                    ral::execution::executor::get_instance()->add_task(
+                            std::move(sampleCacheDatas),
+                            this->output_cache("output_b"),
+                            this,
+                            {{"operation_type", "compute_partition_plan"}});
+
+                    get_samples = false;  // we got enough samples, at least as max_order_by_samples
+                    already_computed_partition_plan = true;
+                }
+            }
+        }
+
+        if(sortedTable){
+            auto num_rows = sortedTable->num_rows();
+            auto num_bytes = sortedTable->sizeInBytes();
+
+            events_logger->info("{ral_id}|{query_id}|{kernel_id}|{input_num_rows}|{input_num_bytes}|{output_num_rows}|{output_num_bytes}|{event_type}|{timestamp_begin}|{timestamp_end}",
+                            "ral_id"_a=context->getNodeIndex(ral::communication::CommunicationData::getInstance().getSelfNode()),
+                            "query_id"_a=context->getContextToken(),
+                            "kernel_id"_a=this->get_id(),
+                            "input_num_rows"_a=num_rows,
+                            "input_num_bytes"_a=num_bytes,
+                            "output_num_rows"_a=num_rows,
+                            "output_num_bytes"_a=num_bytes,
+                            "event_type"_a="compute",
+                            "timestamp_begin"_a=eventTimer.start_time(),
+                            "timestamp_end"_a=eventTimer.end_time());
+        }
+
+        output->addToCache(std::move(sortedTable), "output_a");
     }
-    int max_order_by_samples = 10000;
-    it = config_options.find("MAX_ORDER_BY_SAMPLES_PER_NODE");
+    else if (operation_type == "compute_partition_plan") {
+        avg_bytes_per_row = localTotalNumRows == 0 ? 1 : localTotalBytes/localTotalNumRows;
+        compute_partition_plan(std::move(inputs), avg_bytes_per_row, localTotalNumRows);
+    }
+}
+
+kstatus SortAndSampleKernel::run() {
+    CodeTimer timer;
+
+    std::map<std::string, std::string> config_options = context->getConfigOptions();
+    auto it = config_options.find("MAX_ORDER_BY_SAMPLES_PER_NODE");
     if (it != config_options.end()){
         max_order_by_samples = std::stoi(config_options["MAX_ORDER_BY_SAMPLES_PER_NODE"]);
     }
 
-    bool ordered = false;
-    BatchSequence input(this->input_cache(), this, ordered);
-    std::vector<std::unique_ptr<ral::frame::BlazingTable>> sampledTables;
-    std::vector<ral::frame::BlazingTableView> sampledTableViews;
-    std::size_t localTotalNumRows = 0;
-    std::size_t localTotalBytes = 0;
-    int batch_count = 0;
-    while (input.wait_for_next()) {
-        try {
-            auto batch = input.next();
+    std::unique_ptr <ral::cache::CacheData> cache_data = this->input_cache()->pullCacheData();
+    while (cache_data != nullptr) {
+        std::vector<std::unique_ptr <ral::cache::CacheData> > inputs;
+        inputs.push_back(std::move(cache_data));
 
-            eventTimer.start();
-            auto log_input_num_rows = batch ? batch->num_rows() : 0;
-            auto log_input_num_bytes = batch ? batch->sizeInBytes() : 0;
+        ral::execution::executor::get_instance()->add_task(
+                std::move(inputs),
+                this->output_cache("output_a"),
+                this,
+                {{"operation_type", "ordering_and_get_samples"}});
 
-            auto sortedTable = ral::operators::sort(batch->toBlazingTableView(), this->expression);
-            if (get_samples) {
-                auto sampledTable = ral::operators::sample(batch->toBlazingTableView(), this->expression, order_by_samples_ratio);
-                sampledTableViews.push_back(sampledTable->toBlazingTableView());
-                sampledTables.push_back(std::move(sampledTable));
-            }
-            localTotalNumRows += batch->view().num_rows();
-            localTotalBytes += batch->sizeInBytes();
-
-            // Try samples estimation
-            if(try_num_rows_estimation) {
-                std::tie(estimate_samples, num_rows_estimate) = this->query_graph->get_estimated_input_rows_to_cache(this->get_id(), std::to_string(this->get_id()));
-                population_to_sample = static_cast<uint64_t>(num_rows_estimate * order_by_samples_ratio);
-                population_to_sample = (population_to_sample > max_order_by_samples) ? (max_order_by_samples) : population_to_sample;
-                try_num_rows_estimation = false;
-            }
-            population_sampled += batch->num_rows();
-            if (estimate_samples && population_to_sample > 0 && population_sampled > population_to_sample)	{
-
-                size_t avg_bytes_per_row = localTotalNumRows == 0 ? 1 : localTotalBytes/localTotalNumRows;
-                partition_plan_thread = BlazingThread(&SortAndSampleKernel::compute_partition_plan, this, sampledTableViews, avg_bytes_per_row, num_rows_estimate);
-                estimate_samples = false;
-                get_samples = false;
-            }
-            // End estimation
-
-            eventTimer.stop();
-
-            if(sortedTable){
-                auto log_output_num_rows = sortedTable->num_rows();
-                auto log_output_num_bytes = sortedTable->sizeInBytes();
-
-                events_logger->info("{ral_id}|{query_id}|{kernel_id}|{input_num_rows}|{input_num_bytes}|{output_num_rows}|{output_num_bytes}|{event_type}|{timestamp_begin}|{timestamp_end}",
-                                "ral_id"_a=context->getNodeIndex(ral::communication::CommunicationData::getInstance().getSelfNode()),
-                                "query_id"_a=context->getContextToken(),
-                                "kernel_id"_a=this->get_id(),
-                                "input_num_rows"_a=log_input_num_rows,
-                                "input_num_bytes"_a=log_input_num_bytes,
-                                "output_num_rows"_a=log_output_num_rows,
-                                "output_num_bytes"_a=log_output_num_bytes,
-                                "event_type"_a="compute",
-                                "timestamp_begin"_a=eventTimer.start_time(),
-                                "timestamp_end"_a=eventTimer.end_time());
-            }
-
-            if(this->add_to_output_cache(std::move(sortedTable), "output_a")){
-                batch_count++;
-            }
-
-        } catch(const std::exception& e) {
-            // TODO add retry here
-            // Note that we have to handle the collected samples in a special way. We need to compare to the current batch_count and perhaps evict one set of samples
-            logger->error("{query_id}|{step}|{substep}|{info}|{duration}||||",
-                                        "query_id"_a=context->getContextToken(),
-                                        "step"_a=context->getQueryStep(),
-                                        "substep"_a=context->getQuerySubstep(),
-                                        "info"_a="In SortAndSample kernel batch {} for {}. What: {}"_format(batch_count, expression, e.what()),
-                                        "duration"_a="");
-            throw;
-        }
+        cache_data = this->input_cache()->pullCacheData();
     }
 
-    if (partition_plan_thread.joinable()){
-        partition_plan_thread.join();
-    } else {
-        size_t avg_bytes_per_row = localTotalNumRows == 0 ? 1 : localTotalBytes/localTotalNumRows;
-        compute_partition_plan(sampledTableViews, avg_bytes_per_row, localTotalNumRows);
+    std::unique_lock<std::mutex> lock(kernel_mutex);
+    kernel_cv.wait(lock,[this]{
+        return this->tasks.empty();
+    });
+
+    if (!already_computed_partition_plan) {
+        avg_bytes_per_row = localTotalNumRows == 0 ? 1 : localTotalBytes/localTotalNumRows;
+        compute_partition_plan(std::move(samplesTables), avg_bytes_per_row, localTotalNumRows);
     }
 
     logger->debug("{query_id}|{step}|{substep}|{info}|{duration}|kernel_id|{kernel_id}||",
@@ -317,8 +318,8 @@ PartitionKernel::PartitionKernel(std::size_t kernel_id, const std::string & quer
 }
 
 void PartitionKernel::do_process(std::vector< std::unique_ptr<ral::frame::BlazingTable> > inputs,
-    std::shared_ptr<ral::cache::CacheMachine> output,
-    cudaStream_t stream, const std::map<std::string, std::string>& args) {
+    std::shared_ptr<ral::cache::CacheMachine> /*output*/,
+    cudaStream_t /*stream*/, const std::map<std::string, std::string>& /*args*/) {
     auto & input = inputs[0];
 
     std::vector<ral::distribution::NodeColumnView> partitions = ral::distribution::partitionData(this->context.get(), input->toBlazingTableView(), partitionPlan->toBlazingTableView(), sortColIndices, sortOrderTypes);
@@ -343,14 +344,13 @@ kstatus PartitionKernel::run() {
     std::map<std::string, std::map<int32_t, int> > node_count;
 
     std::tie(sortColIndices, sortOrderTypes, std::ignore) =	ral::operators::get_sort_vars(this->expression);
-    auto& self_node = ral::communication::CommunicationData::getInstance().getSelfNode();
     auto nodes = context->getAllNodes();
 
     // If we have no partitionPlan, its because we have no data, therefore its one partition per node
     num_partitions_per_node = partitionPlan->num_rows() == 0 ? 1 : (partitionPlan->num_rows() + 1) / this->context->getTotalNodes();
 
     std::map<int32_t, int> temp_partitions_map;
-    for (size_t i = 0; i < num_partitions_per_node; i++) {
+    for (int i = 0; i < num_partitions_per_node; i++) {
         temp_partitions_map[i] = 0;
     }
     for (auto &&node : nodes) {
@@ -423,7 +423,7 @@ MergeStreamKernel::MergeStreamKernel(std::size_t kernel_id, const std::string & 
 
 void MergeStreamKernel::do_process(std::vector< std::unique_ptr<ral::frame::BlazingTable> > inputs,
     std::shared_ptr<ral::cache::CacheMachine> output,
-    cudaStream_t stream, const std::map<std::string, std::string>& args) {
+    cudaStream_t /*stream*/, const std::map<std::string, std::string>& /*args*/) {
 
     if (inputs.empty()) {
         // no op
@@ -443,7 +443,7 @@ kstatus MergeStreamKernel::run() {
     CodeTimer timer;
 
     int batch_count = 0;
-    for (auto idx = 0; idx < this->input_.count(); idx++)
+    for (std::size_t idx = 0; idx < this->input_.count(); idx++)
     {
         try {
             auto cache_id = "input_" + std::to_string(idx);
@@ -518,7 +518,7 @@ LimitKernel::LimitKernel(std::size_t kernel_id, const std::string & queryString,
 
 void LimitKernel::do_process(std::vector< std::unique_ptr<ral::frame::BlazingTable> > inputs,
     std::shared_ptr<ral::cache::CacheMachine> output,
-    cudaStream_t stream, const std::map<std::string, std::string>& args) {
+    cudaStream_t /*stream*/, const std::map<std::string, std::string>& /*args*/) {
     CodeTimer eventTimer(false);
     auto & input = inputs[0];
 
@@ -577,10 +577,10 @@ kstatus LimitKernel::run() {
 
         std::vector<std::string> limit_messages_to_wait_for;
         std::vector<std::string> target_ids;
-        for (auto i = 0; i < nodes_to_send.size(); i++) {
-            target_ids.push_back(nodes_to_send[i].id());
+        for (auto & node_to_send : nodes_to_send) {
+            target_ids.push_back(node_to_send.id());
             limit_messages_to_wait_for.push_back(
-                std::to_string(this->context->getContextToken()) + "_" + std::to_string(this->get_id()) + "_" + nodes_to_send[i].id());
+                std::to_string(this->context->getContextToken()) + "_" + std::to_string(this->get_id()) + "_" + node_to_send.id());
         }
         ral::cache::MetadataDictionary extra_metadata;
         extra_metadata.add_value(ral::cache::TOTAL_TABLE_ROWS_METADATA_LABEL, total_batch_rows);
@@ -595,9 +595,9 @@ kstatus LimitKernel::run() {
             extra_metadata);
 
         int64_t prev_total_rows = 0;
-        for (auto i = 0; i < limit_messages_to_wait_for.size(); i++)	{
+        for (std::size_t i = 0; i < limit_messages_to_wait_for.size(); i++) {
             auto meta_message = this->query_graph->get_input_message_cache()->pullCacheData(limit_messages_to_wait_for[i]);
-            if(i < context->getNodeIndex(ral::communication::CommunicationData::getInstance().getSelfNode())){
+            if(static_cast<int>(i) < context->getNodeIndex(ral::communication::CommunicationData::getInstance().getSelfNode())){
                 prev_total_rows += std::stoi(static_cast<ral::cache::GPUCacheDataMetaData*>(meta_message.get())->getMetadata().get_values()[ral::cache::TOTAL_TABLE_ROWS_METADATA_LABEL]);
             }
         }