+++ engine/src/execution_graph/logic_controllers/BatchProcessing.h
@@ -313,9 +313,10 @@ public:
 
 	std::string kernel_name() { return "Output";}
 
-	void do_process(std::vector< std::unique_ptr<ral::frame::BlazingTable> > inputs,
-		std::shared_ptr<ral::cache::CacheMachine> output,
-		cudaStream_t stream, const std::map<std::string, std::string>& args) override {
+	void do_process(std::vector< std::unique_ptr<ral::frame::BlazingTable> > /*inputs*/,
+		std::shared_ptr<ral::cache::CacheMachine> /*output*/,
+		cudaStream_t /*stream*/,
+		const std::map<std::string, std::string>& /*args*/) override {
 			//for now the output kernel is not using do_process
 			//i believe the output should be a cachemachine itself
 			//obviating this concern