+++ engine/src/utilities/CommonOperations.cpp
@@ -191,7 +191,7 @@ void normalize_types(std::unique_ptr<ral::frame::BlazingTable> & table,  const s
 		std::vector<cudf::size_type> column_indices) {
 
 	if (column_indices.size() == 0){
-		RAL_EXPECTS(table->num_columns() == types.size(), "In normalize_types: table->num_columns() != types.size()");
+		RAL_EXPECTS(static_cast<size_t>(table->num_columns()) == types.size(), "In normalize_types: table->num_columns() != types.size()");
 		column_indices.resize(table->num_columns());
 		std::iota(column_indices.begin(), column_indices.end(), 0);
 	} else {