+++ pyblazing/pyblazing/apiv2/context.py
@@ -233,7 +233,6 @@ def initializeBlazing(
     output_cache, input_cache, self_port = cio.initializeCaller(
         ralId,
         worker_id.encode(),
-        0,
         networkInterface.encode(),
         self_port,
         workers_ucp_info,
@@ -288,7 +287,6 @@ def generateGraphs(
     fileTypes,
     ctxToken,
     algebra,
-    accessToken,
     config_options,
     sql,
     single_gpu=False,
@@ -328,7 +326,6 @@ def generateGraphs(
             fileTypes,
             ctxToken,
             algebra,
-            accessToken,
             config_options,
             sql,
         )
@@ -1153,7 +1150,6 @@ def load_config_options_from_env(user_config_options: dict):
         "TABLE_SCAN_KERNEL_NUM_THREADS": 4,
         "MAX_DATA_LOAD_CONCAT_CACHE_BYTE_SIZE": 400000000,
         "FLOW_CONTROL_BYTES_THRESHOLD": 18446744073709551615,  # see https://en.cppreference.com/w/cpp/types/numeric_limits/max
-        "ORDER_BY_SAMPLES_RATIO": 0.1,
         "MAX_ORDER_BY_SAMPLES_PER_NODE": 10000,
         "BLAZING_DEVICE_MEM_CONSUMPTION_THRESHOLD": 0.95,
         "BLAZ_HOST_MEM_CONSUMPTION_THRESHOLD": 0.75,
@@ -1276,10 +1272,6 @@ class BlazingContext(object):
             MAX_DATA_LOAD_CONCAT_CACHE_BYTE_SIZE : The max size in bytes to
                     concatenate the batches read from the scan kernels
                     default: 400000000
-            ORDER_BY_SAMPLES_RATIO : The ratio to multiply the estimated total
-                    number of rows in the SortAndSampleKernel to calculate
-                    the number of samples
-                    default: 0.1
             MAX_ORDER_BY_SAMPLES_PER_NODE : The max number order by samples
                     to capture per node
                     default: 10000
@@ -1823,6 +1815,82 @@ class BlazingContext(object):
             free_memory_dictionary = {}
             free_memory_dictionary[0] = cio.getFreeMemoryCaller()
             return free_memory_dictionary
+
+    def get_max_memory_used(self):
+        """
+        This function returns a dictionary which contains as
+        key the gpuID and as value the max memory (bytes)
+
+        Example
+        --------
+        # single-GPU
+        >>> from blazingsql import BlazingContext
+        >>> bc = BlazingContext()
+        >>> max_mem_used = bc.get_max_memory_used()
+        >>> print(max_mem_used)
+                {0: 4234220154}
+
+        # multi-GPU (4 GPUs):
+        >>> from blazingsql import BlazingContext
+        >>> from dask_cuda import LocalCUDACluster
+        >>> from dask.distributed import Client
+        >>> cluster = LocalCUDACluster()
+        >>> client = Client(cluster)
+        >>> bc = BlazingContext(dask_client=client, network_interface='lo')
+        >>> max_mem_used = bc.get_max_memory_used()
+        >>> print(max_mem_used)
+                {0: 4234220154, 1: 4104210987,
+                 2: 4197720291, 3: 3934320116}
+        """
+        if self.dask_client:
+            dask_futures = []
+            workers_id = []
+            workers = tuple(self.dask_client.scheduler_info()["workers"])
+            for worker_id, worker in enumerate(workers):
+                free_memory = self.dask_client.submit(
+                    cio.getMaxMemoryUsedCaller, workers=[worker], pure=False
+                )
+                dask_futures.append(free_memory)
+                workers_id.append(worker_id)
+            aslist = self.dask_client.gather(dask_futures)
+            free_memory_dictionary = dict(zip(workers_id, aslist))
+            return free_memory_dictionary
+        else:
+            free_memory_dictionary = {}
+            free_memory_dictionary[0] = cio.getMaxMemoryUsedCaller()
+            return free_memory_dictionary
+
+    def reset_max_memory_used(self) -> None:
+        """
+        This function resets the max memory usage counter to 0
+
+        Example
+        --------
+        # single-GPU
+        >>> from blazingsql import BlazingContext
+        >>> bc = BlazingContext()
+        >>> bc.reset_max_memory_used()
+
+        # multi-GPU (4 GPUs):
+        >>> from blazingsql import BlazingContext
+        >>> from dask_cuda import LocalCUDACluster
+        >>> from dask.distributed import Client
+        >>> cluster = LocalCUDACluster()
+        >>> client = Client(cluster)
+        >>> bc = BlazingContext(dask_client=client, network_interface='lo')
+        >>> bc.reset_max_memory_used()
+        """
+        if self.dask_client:
+            dask_futures = []
+            workers = tuple(self.dask_client.scheduler_info()["workers"])
+            for worker_id, worker in enumerate(workers):
+                free_memory = self.dask_client.submit(
+                    cio.resetMaxMemoryUsedCaller, workers=[worker], pure=False
+                )
+                dask_futures.append(free_memory)
+            self.dask_client.gather(dask_futures)
+        else:
+            cio.resetMaxMemoryUsedCaller()
 
     def create_table(self, table_name, input, **kwargs):
         """
@@ -2578,12 +2646,11 @@ class BlazingContext(object):
                     ].tolist()
                     row_group_ids = [row_groups_col[i] for i in row_indices]
                     row_groups_ids.append(row_group_ids)
-        
+
         else:
             actual_files = current_table.files
             uri_values = current_table.uri_values
             row_groups_ids = current_table.row_groups_ids
-        
 
         if self.dask_client is None:
             curr_calcite = current_table.calcite_to_file_indices
@@ -2609,9 +2676,7 @@ class BlazingContext(object):
                     all_sliced_files,
                     all_sliced_uri_values,
                     all_sliced_row_groups_ids,
-                ) = self._sliceRowGroups(
-                    1, actual_files, uri_values, row_groups_ids
-                )
+                ) = self._sliceRowGroups(1, actual_files, uri_values, row_groups_ids)
                 i = 0
                 curr_calcite = current_table.calcite_to_file_indices
                 bt = BlazingTable(
@@ -2670,7 +2735,6 @@ class BlazingContext(object):
                     nodeFilesList.append(bt)
 
         return nodeFilesList
-        
 
     """
     This function has been Deprecated. It is recommended to use ddf.shuffle(on=[colnames])
@@ -2861,7 +2925,6 @@ class BlazingContext(object):
                 nodeList.append(currentTableNodes[j])
 
         ctxToken = random.randint(0, np.iinfo(np.int32).max)
-        accessToken = 0
 
         algebra = get_json_plan(algebra)
 
@@ -2875,7 +2938,6 @@ class BlazingContext(object):
                     fileTypes,
                     ctxToken,
                     algebra,
-                    accessToken,
                     query_config_options,
                     query,
                 )
@@ -2904,7 +2966,6 @@ class BlazingContext(object):
                         fileTypes,
                         ctxToken,
                         algebra,
-                        accessToken,
                         query_config_options,
                         query,
                         single_gpu=True,
@@ -2935,7 +2996,6 @@ class BlazingContext(object):
                             fileTypes,
                             ctxToken,
                             algebra,
-                            accessToken,
                             query_config_options,
                             query,
                             workers=[worker],